Split Dockerfile to stages for dev and prod
We're deploying Matchbox to AWS, and so making changes that support that.

This splits the Dockerfile into stages

- A "base" stage that has the bulk of the the command/config
- A "dev" stage that has configuration for local development
- A "prod" stage that has the configuration for deployment

The dev and prod stages are kept deliberately minimal to try to make sure that
the local dev environment is as close to production as possible. The key
differences between them are that the "prod" stage uses the run mode of fastapi
compared to the "dev" mode in the dev stage, also "prod" does not contain any
environment variables, while "dev" copies them in from the local filesystem -
something we definitely want to avoid when making a production image.