Run the Datadog agent for local dev and pass traces to it
We're improving the monitoring/alerting/debugging setup, and specifically
integrating with Datadog that the department is moving towards.

In order to maintain reusability, Matchbox should continue to work without
Datadog: it is an optional integration, both locally and when deployed.

This change:

1. Runs the Datadog agent when running locally via docker compose using the
   "monitoring" profile

2. Sends traces to it from the api application by wrapping its startup with
   ddtrace-run as the ENTRYPOINT specified in the Dockerfile. It runs the agent
   and ddtrace-run in all cases in order to keep the local environment as similar
   as possible to deployed, but to actually sent data to Datadog, a real
   DD_API_KEY must be set in .datadog-agent-private.env. and the "monitoring"
   profile used with docker compose

Although the Datadog agent would typically not be crucial for local dev work,
the plan is to run the agent when deployed, so this is a stepping stone to
that.